Relocate short command
